fix: rename SignLoginPage
